feat(jira fields): Script to convert text to links in forum threads field.
The new Jira view puts the custom fields in a sidebar, and the urls are not converted to links. This
script finds that field via an adajacent button and converts the strings (urls) to links.

Caveats:
Doesnt work on the "old view" (and doesn't need to).
If Jira changes the way they handle those urls, then this should likely be edited or removed.
There is no logic for handling if they are already links.